# backend/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: MSS Course Platform BFF API
  description: |
    Backend for Frontend API specification for the MSS Course Platform.
    This API defines all endpoints currently implemented using Supabase,
    prepared for migration to Python FastAPI BFF layer.
  version: 1.0.0
  contact:
    name: MSS Course Platform Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-app.vercel.app/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
...
    VideoBase:
      type: object
      required:
        - youtube_url
        - title_ja
      properties:
        youtube_url:
          type: string
          format: uri
          description: Full YouTube URL
        title_ja:
          type: string
          maxLength: 100
        title_en:
          type: string
          maxLength: 100
          nullable: true
        description_ja:
          type: string
          maxLength: 1000
          nullable: true
        description_en:
          type: string
          maxLength: 1000
          nullable: true
    Video:
      allOf:
        - $ref: '#/components/schemas/VideoBase'
        - type: object
          required:
            - id
            - course_id
            - youtube_video_id
            - order_number
          properties:
            id:
              type: string
              format: uuid
            course_id:
              type: string
              format: uuid
            youtube_video_id:
              type: string
              description: Extracted YouTube video ID
            order_number:
              type: integer
              minimum: 1
              description: Display order within course
            duration:
              type: integer
              nullable: true
              description: Video duration in seconds
            duration_seconds:
              type: integer
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    # Progress schemas
...
paths:
...
  /courses/{courseId}/videos:
    get:
      tags:
        - Videos
      summary: List course videos
      description: Get all videos for a course
      operationId: getCourseVideos
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Video'

    post:
      tags:
        - Videos
      summary: Add video to course
      description: Add a new video to course (max 10 videos, owner or admin only)
      operationId: addVideo
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoBase'
      responses:
        '201':
          description: Video added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Video'
        '400':
          description: Invalid input or maximum videos reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /videos/{id}:
    put:
      tags:
        - Videos
      summary: Update video
      description: Update video information
      operationId: updateVideo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoBase'
      responses:
        '200':
          description: Video updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Videos
      summary: Delete video
      description: Delete a video and reorder remaining videos
      operationId: deleteVideo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...